% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartiteD3-package.r
\docType{package}
\name{bipartiteD3-package}
\alias{bipartiteD3}
\alias{bipartiteD3-package}
\title{bipartiteD3: Interactive Bipartite Graphs}
\description{
Generates interactive bipartite graphs using the D3 library. Designed for use with the 'bipartite' analysis package. Sources open source 'viz-js' library Adapted from examples at \url{https://bl.ocks.org/NPashaP} (released under GPL-3).
}
\author{
\strong{Maintainer}: Chris Terry \email{christerry3@btinternet.com}

}
\keyword{internal}
