% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartiteD3-package.r
\docType{package}
\name{bipartiteD3}
\alias{bipartiteD3}
\title{Interactive bipartite graphs}
\description{
This package can generate interactive bipartite graphs through the d3 javascript
library, direct from data formats used in ecology. Please see the vignette for
examples
}
