% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadVisJS.R
\name{LoadVisJS}
\alias{LoadVisJS}
\title{LoadVisJS}
\usage{
LoadVisJS()
}
\description{
Used internally by BP_JS_Writer() and bipartite_D3() to store the necessary JavaScript library
}
\details{
This was originally sourced from the open source vis JavaScript library from
vizjs.org (v1.1.0). However, since this is no longer maintained, the library is hardcoded in here.
}
