% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadVisJS.R
\name{LoadVisJS}
\alias{LoadVisJS}
\title{LoadVisJS}
\usage{
LoadVisJS()
}
\description{
Downloads source code for the open source vis JavaScript library from vizjs.org if it is not already present
in working directory. Uses v1.1.0
}
\details{
Used internally by BP_JS_Writer() and bipartite_D3()
}
